pkp/pkp-lib#6264 DAO adaptations; remove eof and row count calls (no longer supported)

+++ classes/security/AccessKeyDAO.inc.php
@@ -30,15 +30,10 @@ class AccessKeyDAO extends DAO {
 				'SELECT * FROM access_keys WHERE access_key_id = ? AND expiry_date > %s',
 				$this->datetimeToDB(Core::getCurrentDate())
 			),
-			array((int) $accessKeyId)
+			[(int) $accessKeyId]
 		);
-
-		$accessKey = null;
-		if ($result->RecordCount() != 0) {
-			$accessKey =& $this->_returnAccessKeyFromRow($result->GetRowAssoc(false));
-		}
-		$result->Close();
-		return $accessKey;
+		$row = $result->current();
+		return $row ? $this->_returnAccessKeyFromRow((array) $row) : null;
 	}
 
 	/**
@@ -53,15 +48,10 @@ class AccessKeyDAO extends DAO {
 				'SELECT * FROM access_keys WHERE context = ? AND user_id = ? AND expiry_date > %s',
 				$this->datetimeToDB(Core::getCurrentDate())
 			),
-			array($context, $userId)
+			[$context, $userId]
 		);
-
-		$returner = null;
-		if ($result->RecordCount() != 0) {
-			$returner =& $this->_returnAccessKeyFromRow($result->GetRowAssoc(false));
-		}
-		$result->Close();
-		return $returner;
+		$row = $result->current();
+		return $row ? $this->_returnAccessKeyFromRow((array) $row) : null;
 	}
 
 	/**
@@ -82,13 +72,8 @@ class AccessKeyDAO extends DAO {
 			),
 			$paramArray
 		);
-
-		$returner = null;
-		if ($result->RecordCount() != 0) {
-			$returner =& $this->_returnAccessKeyFromRow($result->GetRowAssoc(false));
-		}
-		$result->Close();
-		return $returner;
+		$row = $result->current();
+		return $row ? $this->_returnAccessKeyFromRow((array) $row) : null;
 	}
 
 	/**
@@ -129,12 +114,12 @@ class AccessKeyDAO extends DAO {
 				VALUES
 				(?, %s, ?, ?, ?)',
 				$this->datetimeToDB($accessKey->getExpiryDate())),
-			array(
+			[
 				$accessKey->getKeyHash(),
 				$accessKey->getContext(),
 				$accessKey->getAssocId()==''?null:(int) $accessKey->getAssocId(),
 				(int) $accessKey->getUserId()
-			)
+			]
 		);
 
 		$accessKey->setId($this->getInsertId());
@@ -145,7 +130,7 @@ class AccessKeyDAO extends DAO {
 	 * Update an existing accessKey.
 	 * @param $accessKey AccessKey
 	 */
-	function updateObject(&$accessKey) {
+	function updateObject($accessKey) {
 		return $this->update(
 			sprintf('UPDATE access_keys
 				SET
@@ -156,13 +141,13 @@ class AccessKeyDAO extends DAO {
 					user_id = ?
 				WHERE access_key_id = ?',
 				$this->datetimeToDB($accessKey->getExpiryDate())),
-			array(
+			[
 				$accessKey->getKeyHash(),
 				$accessKey->getContext(),
 				$accessKey->getAssocId()==''?null:(int) $accessKey->getAssocId(),
 				(int) $accessKey->getUserId(),
 				(int) $accessKey->getId()
-			)
+			]
 		);
 	}
 
@@ -170,7 +155,7 @@ class AccessKeyDAO extends DAO {
 	 * Delete an accessKey.
 	 * @param $accessKey AccessKey
 	 */
-	function deleteObject(&$accessKey) {
+	function deleteObject($accessKey) {
 		return $this->deleteAccessKeyById($accessKey->getId());
 	}
 
@@ -179,10 +164,7 @@ class AccessKeyDAO extends DAO {
 	 * @param $accessKeyId int
 	 */
 	function deleteAccessKeyById($accessKeyId) {
-		return $this->update(
-			'DELETE FROM access_keys WHERE access_key_id = ?',
-			array((int) $accessKeyId)
-		);
+		return $this->update('DELETE FROM access_keys WHERE access_key_id = ?', [(int) $accessKeyId]);
 	}
 
 	/**
@@ -193,7 +175,7 @@ class AccessKeyDAO extends DAO {
 	function transferAccessKeys($oldUserId, $newUserId) {
 		return $this->update(
 			'UPDATE access_keys SET user_id = ? WHERE user_id = ?',
-			array((int) $newUserId, (int) $oldUserId)
+			[(int) $newUserId, (int) $oldUserId]
 		);
 	}
